ADM-623:[frontend]feat: fix test
